backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
faiss-cpu==1.12.0
+openai==1.43.0
pypdf==4.2.0
docx2txt==0.8
python-multipart==0.0.9
httpx==0.27.0